fix: Correct weight_gain type in FullyConnectedLayer and adjust gains/biases shape in Decoder

+++ mapping_network.py
@@ -25,7 +25,7 @@ class FullyConnectedLayer(torch.nn.Module):
         self.activation = activation
         self.weight = torch.nn.Parameter(torch.randn([out_features, in_features]) / lr_multiplier)
         self.bias = torch.nn.Parameter(torch.full([out_features], np.float32(bias_init))) if bias else None
-        self.weight_gain = lr_multiplier / np.sqrt(in_features)
+        self.weight_gain = float(lr_multiplier / np.sqrt(in_features))
         self.bias_gain = lr_multiplier
 
     def forward(self, x):
